[문제 해결] 입력하여 바로 검색 기능 관련 문제 해결 * 검색 시 프로그램이 멈추는 문제 해결 * 글자 수 슬라이더의 숫자가 이제는 정수로 나타납니다.

+++ lib/search.dart
@@ -14,7 +14,7 @@ class _SearchPageState extends State<SearchPage> {
   List classList = [];
   bool _isFirst = true;
   bool _liveSearch = true;
-  int _liveSearchCount = 0;
+  double _liveSearchCount = 0.0;
 
   @override
   void initState() {
@@ -46,7 +46,7 @@ class _SearchPageState extends State<SearchPage> {
     dynamic args = ModalRoute.of(context)!.settings.arguments;
     Map rawClassList = args[0][0];
     _liveSearch = args[1] ?? true;
-    _liveSearchCount = args[2] ?? 0;
+    _liveSearchCount = args[2] ?? 0.0;
     if (_isFirst) {
       for (var _dat in rawClassList.values.toList()) {
         for (var _dat2 in _dat) {
